+++ src/http-server/tests/ResponseTest.php
@@ -12,9 +12,11 @@ declare(strict_types=1);
 
 namespace HyperfTest\HttpServer;
 
+use Hyperf\HttpMessage\Cookie\Cookie;
 use Hyperf\HttpMessage\Stream\SwooleStream;
 use Hyperf\HttpMessage\Uri\Uri;
 use Hyperf\HttpServer\Contract\RequestInterface;
+use Hyperf\HttpServer\Contract\ResponseInterface;
 use Hyperf\HttpServer\Response;
 use Hyperf\Utils\ApplicationContext;
 use Hyperf\Utils\Context;
@@ -24,6 +26,7 @@ use Mockery;
 use PHPUnit\Framework\TestCase;
 use Psr\Container\ContainerInterface;
 use Psr\Http\Message\ResponseInterface as PsrResponseInterface;
+use Swoole\Http\Response as SwooleResponse;
 
 /**
  * @internal
@@ -187,5 +190,49 @@ class ResponseTest extends TestCase
         $response = $response->withBody(new SwooleStream('xxx'));
 
         $this->assertInstanceOf(PsrResponseInterface::class, $response);
+        $this->assertInstanceOf(ResponseInterface::class, $response);
+    }
+
+    public function testCookiesAndHeaders()
+    {
+        $container = Mockery::mock(ContainerInterface::class);
+        ApplicationContext::setContainer($container);
+
+        $swooleResponse = Mockery::mock(SwooleResponse::class);
+        $id = uniqid();
+        $cookie1 = new Cookie('Name', 'Hyperf');
+        $cookie2 = new Cookie('Request-Id', $id);
+        $swooleResponse->shouldReceive('status')->with(Mockery::any())->andReturnUsing(function ($code) {
+            $this->assertSame($code, 200);
+        });
+        $swooleResponse->shouldReceive('header')->withAnyArgs()->twice()->andReturnUsing(function ($name, $value) {
+            if ($name == 'X-Token') {
+                $this->assertSame($value, 'xxx');
+            }
+            return true;
+        });
+        $swooleResponse->shouldReceive('rawcookie')->withAnyArgs()->twice()->andReturnUsing(function ($name, $value, ...$args) use ($id) {
+            $this->assertTrue($name == 'Name' || $name == 'Request-Id');
+            $this->assertTrue($value == 'Hyperf' || $value == $id);
+            return true;
+        });
+        $swooleResponse->shouldReceive('end')->once()->andReturn(true);
+
+        Context::set(PsrResponseInterface::class, $psrResponse = new \Hyperf\HttpMessage\Server\Response($swooleResponse));
+
+        $response = new Response();
+        $response = $response->withCookie($cookie1)->withCookie($cookie2)->withHeader('X-Token', 'xxx')->withStatus(200);
+
+        $this->assertInstanceOf(Response::class, $response);
+        $this->assertInstanceOf(ResponseInterface::class, $response);
+
+        $response = $response->raw('Hello Hyperf.');
+        $this->assertNotInstanceOf(Response::class, $response);
+        $this->assertNotInstanceOf(ResponseInterface::class, $response);
+        $this->assertInstanceOf(PsrResponseInterface::class, $response);
+
+        $response->send();
+
+        $this->assertSame($psrResponse, Context::get(PsrResponseInterface::class));
     }
 }